+++ src/components/content-author/author-bio.js
@@ -1,5 +1,5 @@
 import ReactMarkdown from 'react-markdown'
-import { getImageCacheUrl } from 'common/utilities'
+import { getImageCacheUrl } from 'common/utilities/urls/urls'
 import { css } from 'linaria'
 import { breakpointLargeHandset } from 'common/constants'
 import PropTypes from 'prop-types'
