+++ src/components/content-publisher/publisher-attribution.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { css } from 'linaria'
 import { Button } from 'components/buttons/button'
-import { getPublishedDate } from 'common/utilities'
+import { getPublishedDate } from 'common/utilities/date-time/date-time'
 import { breakpointLargeHandset } from 'common/constants'
 import VisibilitySensor from 'components/visibility-sensor/visibility-sensor'
 import { Trans } from 'next-i18next'
