+++ src/connectors/feature-flags/feature-flags.state.js
@@ -1,7 +1,7 @@
 import * as Sentry from '@sentry/nextjs'
 import { put, takeEvery, select } from 'redux-saga/effects'
 import { getUnleashAssignments } from 'common/api/queries/get-unleash-assignments'
-import { arrayToObject } from 'common/utilities'
+import { arrayToObject } from 'common/utilities/object-array/object-array'
 import { FEATURES_HYDRATE, FEATURES_TOGGLE, FEATURES_ASSIGN, FEATURES_SET, HYDRATE } from 'actions'
 import { setCookie } from 'nookies'
